crud 작성 파일 위치 변경 및 엔드포인트 구현
